requirements.txt
Django==4.1.6
djangorestframework==3.14.0
pytz
openai==0.26.5
whitenoise
gunicorn
django-cors-headers
psycopg2
django-allauth==0.51.0
django-extensions==3.2.1
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
django-cors-headers==3.14.0
